renamed files and fixing routes
