+++ src/app/pages/VerifyEntryPage.tsx
@@ -2,9 +2,9 @@
 import { Navbar } from "../components/UI/Navbar"
 // import EntradaAutenticada from '../components/VerificarEntrada/EntradaAutenticada'
 // import EntradaInvalida from '../components/VerificarEntrada/EntradaInvalida'
-import ValidarEntrada from '../components/VerificarEntrada/ValidarEntrada'
+import ValidarEntrada from '../components/VerifyEntry/ValidateEntry'
 
-export default function VerificarEntradaComponent() {
+export default function VerifyEntryPage() {
   // const [entradaValida, setEntradaValida] = useState(true)
 
   return (
